+++ app/src/main/java/com/finsera/ui/fragments/transfer/antar_bank/TransferAntarBankHomeFragment.kt
@@ -1,5 +1,6 @@
 package com.finsera.ui.fragments.transfer.antar_bank
 
+import androidx.fragment.app.viewModels
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -8,6 +9,19 @@ import android.view.ViewGroup
 import com.finsera.R
 
 class TransferAntarBankHomeFragment : Fragment() {
+
+    companion object {
+        fun newInstance() = TransferAntarBankHomeFragment()
+    }
+
+    private val viewModel: TransferAntarBankHomeViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        // TODO: Use the ViewModel
+    }
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
