+++ app/src/main/java/com/finsera/ui/fragments/transfer/antar_bank/TransferAntarbankKonfirmasiFragment.kt
@@ -1,5 +1,6 @@
 package com.finsera.ui.fragments.transfer.antar_bank
 
+import androidx.fragment.app.viewModels
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -7,7 +8,19 @@ import android.view.View
 import android.view.ViewGroup
 import com.finsera.R
 
-class TransferAntarBankFormKonfirmasiFragment : Fragment() {
+class TransferAntarbankKonfirmasiFragment : Fragment() {
+
+    companion object {
+        fun newInstance() = TransferAntarbankKonfirmasiFragment()
+    }
+
+    private val viewModel: TransferAntarbankFormKonfirmasiViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        // TODO: Use the ViewModel
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
