+++ app/src/main/res/layout/fragment_transfer_antar_bank_form.xml
@@ -17,12 +17,14 @@
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:contentDescription="@string/back"
             android:src="@drawable/ic_back_button"
             android:layout_centerVertical="true"/>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/transfer_antarbank"
+            android:contentDescription="@string/transfer_antarbank"
             android:textSize="24sp"
             android:textColor="@android:color/black"
             android:layout_centerInParent="true"/>
@@ -52,19 +54,23 @@
                 <TextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:text="@string/bank_tujuan"
+                    android:text="@string/Pilih_Bank"
+                    android:contentDescription="@string/pilih_bank"
                     android:textSize="16sp"
                     android:textColor="@color/neutral_09" />
-                <EditText
-                    android:id="@+id/banktujuanEditText"
+
+                <LinearLayout
                     android:layout_width="match_parent"
                     android:layout_height="50dp"
-                    android:background="@drawable/edit_text_background_neutral5"
-                    android:gravity="center_vertical"
-                    android:paddingHorizontal="16dp"
                     android:layout_marginTop="10dp"
-                    android:hint="@string/bank_tujuan"
-                    android:singleLine="true" />
+                    android:background="@drawable/edit_text_background_neutral5"
+                    android:gravity="center_vertical">
+                    <Spinner
+                        android:layout_width="match_parent"
+                        android:layout_height="50dp"
+                        android:paddingStart="10dp"
+                        />
+                </LinearLayout>
 
             </LinearLayout>
 
@@ -79,6 +85,7 @@
                 <TextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:contentDescription="@string/nomor_rekening_tujuan"
                     android:text="@string/nomor_rekening_tujuan"
                     android:textSize="16sp"
                     android:textColor="@color/neutral_09" />
@@ -90,6 +97,7 @@
                     android:gravity="center_vertical"
                     android:paddingHorizontal="16dp"
                     android:layout_marginTop="10dp"
+                    android:contentDescription="@string/input_rekening"
                     android:hint="@string/nomor_rekening_tujuan"
                     android:singleLine="true" />
 
@@ -106,6 +114,7 @@
                 <TextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:contentDescription="@string/nominal"
                     android:text="@string/nominal"
                     android:textSize="16sp"
                     android:textColor="@color/neutral_09" />
@@ -117,6 +126,7 @@
                     android:gravity="center_vertical"
                     android:paddingHorizontal="16dp"
                     android:layout_marginTop="10dp"
+                    android:contentDescription="@string/input_nominal"
                     android:hint="@string/nominal"
                     android:singleLine="true" />
 
@@ -133,6 +143,7 @@
                 <TextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:contentDescription="@string/catatan"
                     android:text="@string/catatan"
                     android:textSize="16sp"
                     android:textColor="@color/neutral_09" />
@@ -144,6 +155,7 @@
                     android:paddingHorizontal="16dp"
                     android:layout_marginTop="10dp"
                     android:inputType="text|textMultiLine"
+                    android:contentDescription="@string/input_catatan"
                     android:hint="@string/catatan" />
 
             </LinearLayout>
@@ -160,6 +172,7 @@
                 <TextView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:contentDescription="@string/simpan_ke_daftar_tersimpan_klik_untuk_simpan"
                     android:text="Simpan ke daftar tersimpan"/>
             </LinearLayout>
 
@@ -173,6 +186,7 @@
         android:layout_height="wrap_content"
         android:background="@drawable/button_background_primary"
         app:backgroundTint="@color/primary_blue"
+        android:contentDescription="@string/lanjut"
         android:text="Lanjut"
         android:textSize="20sp"
         android:layout_marginBottom="40dp"
