+++ app/src/main/res/layout/fragment_transfer_antar_bank_form_konfirmasi.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".ui.fragments.transfer.antar_bank.TransferAntarBankFormKonfirmasiFragment">
+    tools:context=".ui.fragments.transfer.antar_bank.TransferAntarbankKonfirmasiFragment">
 
     <RelativeLayout
         app:layout_constraintStart_toStartOf="parent"
@@ -18,6 +18,7 @@
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:contentDescription="@string/back"
             android:src="@drawable/ic_back_button"
             android:layout_centerVertical="true"/>
 
@@ -25,6 +26,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
+            android:contentDescription="@string/konfirmasi_transaksi"
             android:text="@string/konfirmasi_transaksi"
             android:textColor="@color/neutral_09"
             android:textSize="26sp" />
@@ -223,6 +225,7 @@
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:contentDescription="@string/masukkan_pin_pembayaran"
             android:text="@string/masukkan_pin_pembayaran"
             android:textSize="16sp"
             android:textColor="@color/neutral_09"
@@ -246,7 +249,7 @@
                 android:padding="16dp"
                 android:background="@drawable/edit_text_background_neutral5"
                 android:hint="@string/tv_hint_password_login"
-                android:contentDescription="@string/tv_talkback_password_login"
+                android:contentDescription="@string/input_pin"
                 android:textSize="@dimen/heading_5"
                 android:inputType="numberDecimal" />
 
@@ -258,6 +261,7 @@
         android:layout_height="wrap_content"
         android:background="@drawable/button_background_primary"
         app:backgroundTint="@color/primary_blue"
+        android:contentDescription="@string/lanjut"
         android:text="Lanjut"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
