--- app/src/main/res/layout/fragment_transfer_antarbank_form_konfirmasi.xml
@@ -1,144 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.fragments.transfer.antar_bank.TransferAntarbankFormKonfirmasi"
-    android:paddingHorizontal="16dp">
-
-    <RelativeLayout
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:id="@+id/topmenu"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="32dp">
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_back_button"
-            android:layout_centerVertical="true"/>
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/konfirmasi_transaksi"
-            android:textSize="24sp"
-            android:textColor="@color/neutral_09"
-            android:layout_centerInParent="true"/>
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/topmenu"
-        android:layout_marginTop="48dp"
-        android:id="@+id/layoutbanktujuan">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/bank_tujuan"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:text="@string/bank_tujuan"
-            android:id="@+id/tvBankTujuan"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/layoutbanktujuan"
-        android:layout_marginTop="24dp"
-        android:id="@+id/layoutnomorrekening">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/rekening_tujuan"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:text="1234567890"
-            android:id="@+id/tvRekeningTujuan"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/layoutnomorrekening"
-        android:layout_marginTop="24dp"
-        android:id="@+id/layoutnamapenerima">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Nama Penerima"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:text="Budhi Darma"
-            android:id="@+id/tvNamaPenerima"
-            android:textColor="@color/neutral_09"
-            android:textSize="16sp"/>
-
-    </RelativeLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:id="@+id/layoutmasukkanpin"
-        app:layout_constraintTop_toBottomOf="@id/layoutnamapenerima"
-        android:layout_marginTop="16dp">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/masukkan_pin_pembayaran"
-            android:textSize="16sp"
-            android:textColor="@color/neutral_09"
-            android:layout_marginTop="32dp"/>
-        <EditText
-            android:id="@+id/nominalEditText"
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:background="@drawable/edit_text_background_neutral5"
-            android:drawableEnd="@drawable/ic_visibility_off"
-            android:gravity="center_vertical"
-            android:paddingHorizontal="16dp"
-            android:layout_marginTop="10dp"
-            android:hint="@string/pin"
-            android:singleLine="true" />
-
-    </LinearLayout>
-    <com.google.android.material.button.MaterialButton
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/button_background_primary"
-        app:backgroundTint="@color/primary_blue"
-        android:text="Lanjut"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginBottom="100dp"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
