+++ app/src/main/res/layout/fragment_transfer_virtual_account_form_konfirmasi.xml
@@ -17,6 +17,7 @@
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:contentDescription="@string/back"
             android:src="@drawable/ic_back_button"
             android:layout_centerVertical="true"/>
 
@@ -24,6 +25,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
+            android:contentDescription="@string/konfirmasi_transaksi"
             android:text="@string/konfirmasi_transaksi"
             android:textColor="@color/neutral_09"
             android:textSize="26sp" />
@@ -170,6 +172,7 @@
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:contentDescription="@string/masukkan_pin_pembayaran"
             android:text="@string/masukkan_pin_pembayaran"
             android:textSize="16sp"
             android:textColor="@color/neutral_09"
@@ -192,8 +195,8 @@
                 android:layout_height="wrap_content"
                 android:padding="16dp"
                 android:background="@drawable/edit_text_background_neutral5"
-                android:hint="@string/tv_hint_password_login"
-                android:contentDescription="@string/tv_talkback_password_login"
+                android:hint="@string/pin"
+                android:contentDescription="@string/input_pin"
                 android:textSize="@dimen/heading_5"
                 android:inputType="numberDecimal" />
 
@@ -205,6 +208,7 @@
         android:layout_height="wrap_content"
         android:background="@drawable/button_background_primary"
         app:backgroundTint="@color/primary_blue"
+        android:contentDescription="@string/lanjut"
         android:text="Lanjut"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
