+++ app/src/main/res/layout/fragment_transfer_virtual_account_home.xml
@@ -16,13 +16,14 @@
         app:layout_constraintTop_toTopOf="@id/tv_title_label"
         app:layout_constraintBottom_toBottomOf="@id/tv_title_label"
         android:background="?attr/selectableItemBackgroundBorderless"
-        android:contentDescription="@string/back_button" />
+        android:contentDescription="@string/back" />
 
     <TextView
         android:id="@+id/tv_title_label"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Virtual Account"
+        android:contentDescription="@string/virtual_account"
+        android:text="@string/virtual_account"
         android:textSize="@dimen/heading_3"
         android:textColor="@android:color/black"
         android:layout_marginTop="16dp"
@@ -37,6 +38,7 @@
         android:layout_marginTop="16dp"
         android:backgroundTint="@color/primary_blue"
         android:padding="10dp"
+        android:contentDescription="@string/daftar_tersimpan"
         android:text="@string/daftar_tersimpan"
         android:textSize="@dimen/button_1"
         app:cornerRadius="8dp"
@@ -51,6 +53,7 @@
         android:layout_marginStart="10dp"
         android:backgroundTint="@color/neutral_03"
         android:padding="10dp"
+        android:contentDescription="@string/input_baru_klik_untuk_transfer"
         android:text="@string/input_baru"
         android:textColor="@color/neutral_02"
         android:textSize="@dimen/button_1"
@@ -80,6 +83,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@drawable/et_textfield_background_2"
+            android:contentDescription="@string/cari_daftar_tersimpan"
             android:hint="@string/cari_daftar_tersimpan"
             android:inputType="text"
             android:padding="12dp"
