+++ app/src/main/res/values/strings.xml
@@ -16,12 +16,12 @@
     <string name="rekening_tujuan">Rekening Tujuan</string>
     <string name="pin">PIN</string>
     <string name="masukkan_pin_pembayaran">Masukkan PIN Pembayaran</string>
-<<<<<<< HEAD
-=======
+
+
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="e_wallet">E-Wallet</string>
-    <string name="tambah_penerima_ke_daftar_tersimpan">Tambah penerima ke daftar tersimpan agar \ntransaksi ke orang terdekat lebih mudah</string>
+    <string name="tambah_penerima_ke_daftar_tersimpan">Tambah penerima ke daftar tersimpan agar transaksi ke orang terdekat lebih mudah</string>
     <string name="empty_state_image">Empty List Image</string>
     <string name="back_button">back button</string>
     <string name="logo_gopay">logo gopay</string>
@@ -109,5 +109,23 @@
     <string name="tv_navbar_title_qris">QRIS</string>
     <string name="tv_navbar_title_favorit">Favorit</string>
     <string name="tv_navbar_title_akun">Akun</string>
->>>>>>> af5c3d76d44512ee7d852f96c782cd64ed3dcfca
+    <string name="pilih_bank">pilih bank</string>
+    <string name="input_rekening">input rekening</string>
+    <string name="input_nominal">input nominal</string>
+    <string name="input_catatan">input catatan</string>
+    <string name="simpan_ke_daftar_tersimpan_klik_untuk_simpan">simpan ke daftar tersimpan klik untuk simpan</string>
+    <string name="input_pin">input pin</string>
+    <string name="input_baru_klik_untuk_transfer">input baru klik untuk transfer</string>
+    <string name="belum_ada_daftar_tersimpan_empty">Belum ada daftar tersimpan tambah penerima ke daftar tersimpan agar transaksi ke orang terdekat lebih mudah</string>
+    <string name="transfer_sesama_bank">Transfer Sesama Bank</string>
+    <string name="Pilih_Bank">Pilih Bank</string>
+    <string name="trouble">Oops! Sepertinya ada masalah</string>
+    <string name="network_error_message">Pastikan perangkat anda terhubung pada jaringan. Tunggu beberapa saat dan coba lagi buka aplikasi.</string>
+    <string name="failed_transfer">oops, sepertinya ada masalah pastikan perangkat anda terhubung pada jaringan. tunggu beberapa saat dan coba lagi buka aplikasi</string>
+    <string name="transfer_virtual_account">Transfer Virtual Account</string>
+    <string name="nomor_va">Nomor Virtual Account</string>
+    <string name="input_nomor_va">input nomor virtual account</string>
+    <string name="nomor_VA">Nomor VA</string>
+    <string name="virtual_account">Virtual Account</string>
+
 </resources>
